adjustments to viewmodel again
